refactor(api): move to utils
